requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
binaryornot==0.4.4
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.6.0
cryptography==43.0.3
daphne==4.1.2
Django==4.2.16
django-allauth==65.0.2
django-anymail==11.1
django-browser-reload==1.15.0
django-environ==0.11.2
django-formtools==2.5.1
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
packaging==24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.2
tornado==6.4.1
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.1
